refactor: have a better structure for all things ai viz
