+++ packages/common/package.json
@@ -25,6 +25,7 @@
         "cronstrue": "^2.23.0",
         "dayjs": "^1.11.9",
         "dependency-graph": "^0.11.0",
+        "echarts": "^5.6.0",
         "fast-json-patch": "^3.1.1",
         "handlebars": "^4.7.8",
         "js-yaml": "^4.1.0",
