+++ packages/common/src/ee/AiAgent/chartConfig/slack/index.ts
@@ -1,0 +1,4 @@
+export {
+    getSlackAiEchartsConfig,
+    type SlackAiToolArgs,
+} from './getSlackAiEchartsConfig';
