+++ packages/common/src/ee/AiAgent/chartConfig/web/generateTableVizConfigTool/getTableChartConfig.ts
@@ -1,0 +1,8 @@
+import {
+    ChartType,
+    type TableChartConfig,
+} from '../../../../../types/savedCharts';
+
+export const getTableChartConfig = (): TableChartConfig => ({
+    type: ChartType.TABLE,
+});
