+++ packages/common/src/ee/AiAgent/chartConfig/web/index.ts
@@ -1,0 +1,9 @@
+// runQuery tool
+export { getWebAiChartConfig } from './getWebAiChartConfig';
+
+// shared utils
+export { getAvailableChartTypes } from './shared/getAvailableChartTypes';
+export { getGroupByDimensions } from './shared/getGroupByDimensions';
+
+// Types
+export type { AiAgentChartTypeOption, AiVizMetadata } from './types';
