+++ packages/common/src/ee/AiAgent/chartConfig/web/shared/formatPivotValueLabel.ts
@@ -1,0 +1,26 @@
+import {
+    friendlyName,
+    isField,
+    type ItemsMap,
+} from '../../../../../types/field';
+import { type PivotReference } from '../../../../../types/savedCharts';
+import { formatItemValue } from '../../../../../utils/formatting';
+
+export const formatPivotValueLabel = (
+    pivotReference: PivotReference,
+    fieldsMap: ItemsMap,
+): string => {
+    if (!pivotReference.pivotValues?.[0]) {
+        return '';
+    }
+
+    const pivotFieldId = pivotReference.pivotValues[0].field;
+    const pivotField = fieldsMap[pivotFieldId];
+    const pivotValue = pivotReference.pivotValues[0].value;
+
+    if (pivotField && isField(pivotField)) {
+        return formatItemValue(pivotField, pivotValue);
+    }
+
+    return friendlyName(String(pivotValue));
+};
