+++ packages/common/src/ee/AiAgent/chartConfig/web/types.ts
@@ -1,0 +1,13 @@
+export type AiAgentChartTypeOption =
+    | 'table'
+    | 'bar'
+    | 'horizontal'
+    | 'line'
+    | 'scatter'
+    | 'pie'
+    | 'funnel';
+
+export type AiVizMetadata = {
+    title?: string;
+    description?: string;
+};
