+++ packages/frontend/src/ee/features/aiCopilot/components/ChatElements/AgentVisualizationChartTypeSwitcher.tsx
@@ -1,6 +1,6 @@
 import {
     getAvailableChartTypes,
-    type ChartTypeOption,
+    type AiAgentChartTypeOption,
     type MetricQuery,
 } from '@lightdash/common';
 import { SegmentedControl } from '@mantine-8/core';
@@ -17,12 +17,12 @@ import MantineIcon from '../../../../../components/common/MantineIcon';
 
 type Props = {
     metricQuery: MetricQuery;
-    selectedChartType: ChartTypeOption;
-    onChartTypeChange: (chartType: ChartTypeOption) => void;
+    selectedChartType: AiAgentChartTypeOption;
+    onChartTypeChange: (chartType: AiAgentChartTypeOption) => void;
     hasGroupByDimensions: boolean;
 };
 
-const CHART_TYPE_ICONS: Record<ChartTypeOption, typeof IconTable> = {
+const CHART_TYPE_ICONS: Record<AiAgentChartTypeOption, typeof IconTable> = {
     table: IconTable,
     bar: IconChartBar,
     horizontal: IconChartBar,
@@ -48,7 +48,9 @@ export const AgentVisualizationChartTypeSwitcher: FC<Props> = ({
     return (
         <SegmentedControl
             value={selectedChartType}
-            onChange={(value) => onChartTypeChange(value as ChartTypeOption)}
+            onChange={(value) =>
+                onChartTypeChange(value as AiAgentChartTypeOption)
+            }
             data={availableChartTypes
                 .filter((chartType) => {
                     // Pie and funnel charts are not supported with group by dimensions, they're meant to be used with a single dimension
