+++ packages/frontend/src/ee/features/aiCopilot/components/ChatElements/AiVisualizationRenderer.tsx
@@ -1,11 +1,12 @@
 import {
     AiResultType,
-    CartesianSeriesType,
     ChartType,
     ECHARTS_DEFAULT_COLORS,
+    getGroupByDimensions,
+    getWebAiChartConfig,
+    type AiAgentChartTypeOption,
     type ApiAiAgentThreadMessageVizQuery,
     type ApiError,
-    type ChartTypeOption,
     type ToolRunQueryArgs,
     type ToolTableVizArgs,
     type ToolTimeSeriesArgs,
@@ -24,16 +25,12 @@ import MetricQueryDataProvider from '../../../../../components/MetricQueryData/M
 import UnderlyingDataModal from '../../../../../components/MetricQueryData/UnderlyingDataModal';
 import { type EchartSeriesClickEvent } from '../../../../../components/SimpleChart';
 import ErrorBoundary from '../../../../../features/errorBoundary/ErrorBoundary';
-import {
-    getExpectedSeriesMap,
-    mergeExistingAndExpectedSeries,
-} from '../../../../../hooks/cartesianChartConfig/utils';
 import { type EChartSeries } from '../../../../../hooks/echarts/useEchartsCartesianConfig';
 import useHealth from '../../../../../hooks/health/useHealth';
 import { useOrganization } from '../../../../../hooks/organization/useOrganization';
 import { useExplore } from '../../../../../hooks/useExplore';
 import { type InfiniteQueryResults } from '../../../../../hooks/useQueryResults';
-import { getChartConfigFromAiAgentVizConfig } from '../../utils/echarts';
+import { getComputedSeries } from '../../utils/aiVisualizationRenderer/getComputedSeries';
 import { AgentVisualizationChartTypeSwitcher } from './AgentVisualizationChartTypeSwitcher';
 import AgentVisualizationFilters from './AgentVisualizationFilters';
 import AgentVisualizationMetricsAndDimensions from './AgentVisualizationMetricsAndDimensions';
@@ -67,7 +64,7 @@ export const AiVisualizationRenderer: FC<Props> = ({
         useState<EchartSeriesClickEvent | null>(null);
     const [echartSeries, setEchartSeries] = useState<EChartSeries[]>([]);
     const [selectedChartType, setSelectedChartType] =
-        useState<ChartTypeOption | null>(null);
+        useState<AiAgentChartTypeOption | null>(null);
 
     const resultsData = useMemo(
         () => ({
@@ -78,76 +75,28 @@ export const AiVisualizationRenderer: FC<Props> = ({
         [results, metricQuery, fields],
     );
 
-    const {
-        echartsConfig,
-        type: aiResultType,
-        vizTool,
-    } = useMemo(() => {
-        return getChartConfigFromAiAgentVizConfig({
-            vizConfig: chartConfig,
+    const chartConfigFromAiAgentVizConfig = useMemo(
+        () =>
+            getWebAiChartConfig({
+                vizConfig: chartConfig,
+                metricQuery,
+                maxQueryLimit: health?.query.maxLimit,
+                fieldsMap: fields,
+                overrideChartType: selectedChartType ?? undefined,
+            }),
+        [
+            chartConfig,
             metricQuery,
-            maxQueryLimit: health?.query.maxLimit,
-            fieldsMap: fields,
-            overrideChartType: selectedChartType ?? undefined,
-        });
-    }, [
-        chartConfig,
-        metricQuery,
-        health?.query.maxLimit,
-        fields,
-        selectedChartType,
-    ]);
+            health?.query.maxLimit,
+            fields,
+            selectedChartType,
+        ],
+    );
 
-    const groupByDimensions: string[] | undefined = useMemo(() => {
-        switch (aiResultType) {
-            case AiResultType.QUERY_RESULT:
-                return vizTool.chartConfig?.groupBy ?? undefined;
-            case AiResultType.VERTICAL_BAR_RESULT:
-            case AiResultType.TIME_SERIES_RESULT:
-                return vizTool.vizConfig.breakdownByDimension
-                    ? [vizTool.vizConfig.breakdownByDimension]
-                    : undefined;
-        }
-
-        return undefined;
-    }, [vizTool, aiResultType]);
-
-    const computedSeries = useMemo(() => {
-        if (aiResultType !== AiResultType.QUERY_RESULT) return [];
-        if (echartsConfig.type !== ChartType.CARTESIAN) return [];
-        if (!echartsConfig.config?.eChartsConfig?.series?.length) return [];
-        if (!echartsConfig.config?.layout?.xField) return [];
-        if (!echartsConfig.config?.layout?.yField) return [];
-
-        const firstSerie = echartsConfig.config?.eChartsConfig?.series?.[0];
-
-        const expectedSeriesMap = getExpectedSeriesMap({
-            defaultSmooth: firstSerie?.smooth,
-            defaultShowSymbol: firstSerie?.showSymbol,
-            defaultAreaStyle: firstSerie?.areaStyle,
-            defaultCartesianType: CartesianSeriesType.BAR,
-            availableDimensions: metricQuery.dimensions,
-            isStacked: false,
-            pivotKeys: groupByDimensions,
-            resultsData: resultsData,
-            xField: echartsConfig.config.layout.xField,
-            yFields: echartsConfig.config.layout.yField,
-            defaultLabel: firstSerie?.label,
-            itemsMap: fields,
-        });
-        const newSeries = mergeExistingAndExpectedSeries({
-            expectedSeriesMap,
-            existingSeries: echartsConfig.config?.eChartsConfig?.series || [],
-        });
-        return newSeries;
-    }, [
-        echartsConfig,
-        metricQuery,
-        fields,
-        groupByDimensions,
-        resultsData,
-        aiResultType,
-    ]);
+    const groupByDimensions: string[] | undefined = useMemo(
+        () => getGroupByDimensions(chartConfigFromAiAgentVizConfig),
+        [chartConfigFromAiAgentVizConfig],
+    );
 
     const displayMetricsAndDimensions = useMemo(
         () =>
@@ -162,7 +111,7 @@ export const AiVisualizationRenderer: FC<Props> = ({
             ? chartConfig.chartConfig?.defaultVizType ?? 'table'
             : 'table';
 
-    if (!echartsConfig) {
+    if (!chartConfigFromAiAgentVizConfig.echartsConfig) {
         return (
             <Center h={300}>
                 <Stack gap="xs" align="center">
@@ -185,7 +134,7 @@ export const AiVisualizationRenderer: FC<Props> = ({
             <VisualizationProvider
                 key={selectedChartType ?? 'default'}
                 resultsData={resultsData}
-                chartConfig={echartsConfig}
+                chartConfig={chartConfigFromAiAgentVizConfig.echartsConfig}
                 parameters={
                     queryExecutionHandle.data.query.usedParametersValues
                 }
@@ -209,11 +158,20 @@ export const AiVisualizationRenderer: FC<Props> = ({
                     setEchartsClickEvent(e);
                     setEchartSeries(series);
                 }}
-                computedSeries={computedSeries}
+                computedSeries={getComputedSeries({
+                    aiResultType: chartConfigFromAiAgentVizConfig.type,
+                    echartsConfig:
+                        chartConfigFromAiAgentVizConfig.echartsConfig,
+                    metricQuery,
+                    groupByDimensions,
+                    resultsData,
+                    fields,
+                })}
             >
                 <Stack gap="md" h="100%">
                     {headerContent && headerContent}
-                    {aiResultType === AiResultType.QUERY_RESULT && (
+                    {chartConfigFromAiAgentVizConfig.type ===
+                        AiResultType.QUERY_RESULT && (
                         <Group justify="flex-end">
                             <AgentVisualizationChartTypeSwitcher
                                 metricQuery={metricQuery}
@@ -239,7 +197,8 @@ export const AiVisualizationRenderer: FC<Props> = ({
                             data-testid="ai-visualization"
                         />
 
-                        {echartsConfig.type === ChartType.CARTESIAN && (
+                        {chartConfigFromAiAgentVizConfig.echartsConfig.type ===
+                            ChartType.CARTESIAN && (
                             <SeriesContextMenu
                                 echartSeriesClickEvent={
                                     echartsClickEvent ?? undefined
