+++ packages/frontend/src/ee/features/aiCopilot/utils/dashboardChartConverter.ts
@@ -1,9 +1,9 @@
 import {
+    getWebAiChartConfig,
     type ApiAiAgentThreadMessageVizQuery,
     type CreateSavedChartVersion,
     type DashboardVisualization,
 } from '@lightdash/common';
-import { getChartConfigFromAiAgentVizConfig } from './echarts';
 
 function convertAiVisualizationToCreateSavedChartVersion(
     aiVizData: ApiAiAgentThreadMessageVizQuery,
@@ -20,7 +20,7 @@ function convertAiVisualizationToCreateSavedChartVersion(
     const { query, metadata } = aiVizData;
     const { metricQuery } = query;
 
-    const chartConfig = getChartConfigFromAiAgentVizConfig({
+    const chartConfig = getWebAiChartConfig({
         vizConfig: dashboardVisualization,
         metricQuery,
         maxQueryLimit: options.maxQueryLimit,
